// src/core/settings/settings.interface.tsx
export enum ConfigurationMode {
    SIMPLE = 'simple',
    ADVANCE = 'advanced'
}
export interface IConfiguratioPropertiesPool {
    hostname?: string;
    port?: number;
    username?: string;
    password?: string;
    sslEnabled?: boolean;
}

export enum RandomXMode {
    AUTO = 'auto',
    FAST = 'fast',
    LIGHT = 'light'
}